+++ app/javascript/src/components/Dashboard/Notes/DeleteAlert.jsx
@@ -21,7 +21,6 @@ const DeleteAlert = ({
       refetch();
     } catch (error) {
       logger.error(error);
-    } finally {
       setDeleting(false);
     }
   };
