+++ app/javascript/src/components/Dashboard/Notes/Pane/Form.jsx
@@ -10,9 +10,9 @@ import { NOTES_FORM_VALIDATION_SCHEMA } from "../constants";
 
 export default function NoteForm({ onClose, refetch, note, isEdit }) {
   const [submitted, setSubmitted] = useState(false);
+
   const handleSubmit = async values => {
     try {
-      setSubmitted(true);
       if (isEdit) {
         await notesApi.update(note.id, values);
       } else {
@@ -33,7 +33,7 @@ export default function NoteForm({ onClose, refetch, note, isEdit }) {
       validateOnChange={submitted}
       validationSchema={NOTES_FORM_VALIDATION_SCHEMA}
     >
-      {({ isSubmitting, handleSubmit }) => (
+      {({ isSubmitting }) => (
         <Form className="w-full">
           <Pane.Body className="space-y-6">
             <Input
@@ -59,11 +59,7 @@ export default function NoteForm({ onClose, refetch, note, isEdit }) {
               className="mr-3"
               disabled={isSubmitting}
               loading={isSubmitting}
-              onClick={e => {
-                e.preventDefault();
-                setSubmitted(true);
-                handleSubmit();
-              }}
+              onClick={() => setSubmitted(true)}
             />
             <Button
               onClick={onClose}
